create many files for candidate

+++ src/main/java/by/team34/entity/CandidateExperience.java
@@ -25,7 +25,7 @@ public class CandidateExperience {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "ID")
-    private int id;
+    private Long id;
 
     @ManyToOne
     @JoinColumn(name = "CANDIDATE_ID")
@@ -52,11 +52,11 @@ public class CandidateExperience {
             inverseJoinColumns = @JoinColumn(name = "RESPONSIBILITY"))
     private Set<Responsibility> responsibilities;
 
-    public final int getId() {
+    public final Long getId() {
         return id;
     }
 
-    public final void setId(final int id) {
+    public final void setId(final Long id) {
         this.id = id;
     }
